+++ .eslintrc.js
@@ -9,21 +9,13 @@ module.exports = {
 		mocha: true
 	},
 
-	globals: {
-		__rootpath: true,
-		coreRequire: true,
-		mainRequire: true,
-		JANIS_CORE: true,
-		JANIS_ENV: true,
-		JANIS_ENV_ALIAS: true
+	settings: {
+		'import/core-modules': ['@aws-sdk/client-sns']
 	},
 
 	parserOptions: {
-		sourceType: 'script'
-	},
-
-	settings: {
-		'import/core-modules': ['aws-sdk', 'lodash', 'yamljs', 'openapi-schema-validator']
+		sourceType: 'script',
+		ecmaVersion: 2024
 	},
 
 	rules: {
@@ -36,6 +28,7 @@ module.exports = {
 		'consistent-return': 0,
 		'prefer-template': 0,
 		'import/no-unresolved': 0,
+		'import/extensions': 0,
 		'import/no-extraneous-dependencies': ['error', { devDependencies: ['**/tests/**/*.js'] }],
 
 		'no-bitwise': 0,
@@ -117,6 +110,7 @@ module.exports = {
 			ObjectExpression: { minProperties: 5, multiline: true, consistent: true },
 			ObjectPattern: { minProperties: 5, multiline: true, consistent: true }
 		}],
-		'nonblock-statement-body-position': ['error', 'below', { overrides: { else: 'any' } }]
+		'nonblock-statement-body-position': ['error', 'below', { overrides: { else: 'any' } }],
+		strict: ['error', 'global']
 	}
 };